Fixed some style-bugs, rewrote some style in email-notification, added several additional checks for arrays params

+++ src/1.3/backup.sh
@@ -67,7 +67,7 @@ then
   exclusionList[0]="--exclude=*.git*";
 
   # Array of project names for backup
-  backupProjectName[0]="my_project";
+  backupProjectName[0]="this_is_my_project";
 
   # If you want to back up one database, you must specify its name, and set the 'ALL_DATA_BASE' parameter to 'no'
   dataBaseName[0]="site";
@@ -88,7 +88,7 @@ create_backup()
       # Create a directory for the archive, if not created
       mkdir -p ${BACKUP_ROOT_DIR}
 
-      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Dir ${BACKUP_ROOT_DIR} isn't exist and was created.";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Dir ${BACKUP_ROOT_DIR} isn't exist and was created.";
     fi;
     #
     if echo ${MIN_FREE_SPACE} | awk 'match($0, /[0-9]+MB/) { print substr( $0, RSTART, RLENGTH )}';
@@ -104,7 +104,7 @@ create_backup()
     #
     if [ "${MIN_FREE_SPACE}" -gt "0" ];
     then
-      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check the amount of free space on the HDD";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check the amount of free space on the HDD";
 
       # Check the amount of free space on the HDD
       freespace=$(df -m ${BACKUP_ROOT_DIR} | grep dev | awk '{print $4}');
@@ -112,16 +112,16 @@ create_backup()
       # Check for free space on the HDD
       if [ "${MIN_FREE_SPACE}" -lt "${freespace}" ];
       then
-      	message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r The free space on the hard drive is over. Clear old archives.";
+      	message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - The free space on the hard drive is over. Clear old archives.";
 
         #echo "The free space on the hard drive is over. Clear old archives."
 
         # Delete old archives, with minimum quantity verification
-        message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Delete old archives, with minimum quantity verification";
+        message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Delete old archives, with minimum quantity verification";
 
         clean_by_count
 
-        message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Continue to backup";
+        message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Continue to backup";
         #echo "Continue to backup"
       fi;
     fi;
@@ -135,39 +135,43 @@ create_backup()
     # Create a backup of the specified directory in the directory with the archive
     if [ "${projects_counter}" -gt "0" ]
     then
-      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check whether you need to back up the file system";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check whether you need to back up the file system";
 
       # Check whether you need to back up the file system
       if [ "${FILESYSTEM_BACKUP}" = "yes" ]
       then
-        while [ "$i" != "${projects_counter}" ]; do
-          # The full path of the backup directory
-          pathway=${BACKUP_ROOT_DIR}"/"${backupProjectName[$i]}"/"${date_archived};
-
-          # Check the existence of the backup directory
-          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check the existence of the backup directory";
-
-          if [ ! -d "${pathway}" ]
+        while [ "$i" != "${projects_counter}" ];
+        do
+          if [ -n ${backupProjectName[$i]} ];
           then
-            # Create a directory for the archive, if it was not created earlier
-            message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Create a directory for the archive, if it was not created earlier";
-            mkdir -p ${pathway}
+            # The full path of the backup directory
+            pathway="$BACKUP_ROOT_DIR/${backupProjectName[$i]}/$date_archived";
+
+            # Check the existence of the backup directory
+            message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check the existence of the backup directory";
+
+            if [ ! -d "${pathway}" ]
+            then
+              # Create a directory for the archive, if it was not created earlier
+              message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Create a directory for the archive, if it was not created earlier";
+              mkdir -p ${pathway}
+            fi;
+
+            # Check if we have any exceptions for archiving
+            message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check if we have any exceptions for archiving";
+
+            if [ "${exclusions_counter}" -gt "0" ]
+            then
+              zip -9 -r "$pathway/${backupProjectName[$i]}.zip ${backupProjectDir[$i]} ${exclusionList[$i]}"
+            elif [ "${exclusions_counter}" -eq "0" ]
+            then
+              zip -9 -r "$pathway/${backupProjectName[$i]}.zip ${backupProjectDir[$i]}"
+            fi;
+
+            message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Project archiving ${backupProjectName[$i]} is completed";
+
+            i=$(( i + 1 ))
           fi;
-
-          # Check if we have any exceptions for archiving
-          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check if we have any exceptions for archiving";
-
-          if [ "${exclusions_counter}" -gt "0" ]
-          then
-            zip -9 -r ${pathway}"/"${backupProjectName[$i]}.zip ${backupProjectDir[$i]} ${exclusionList[$i]}
-          elif [ "${exclusions_counter}" -eq "0" ]
-          then
-            zip -9 -r ${pathway}"/"${backupProjectName[$i]}.zip ${backupProjectDir[$i]}
-          fi;
-
-          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Project archiving ${backupProjectName[$i]} is completed";
-
-          i=$(( i + 1 ))
         done
       fi;
     fi;
@@ -176,31 +180,31 @@ create_backup()
     db_user_counter=${#dataBaseLogin[@]}
 
     # Check whether MySQL databases are needed
-    message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check whether MySQL databases are needed";
+    message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check whether MySQL databases are needed";
 
     if [ "${MYSQL_BACKUP}" = "yes" ]
     then
       while [ "$u" != "${db_user_counter}" ]; do
-        if [ -n "${dataBaseLogin[$u]}" ] && [ -n ${dataBasePassword[$u]} ]
+        if [ -n "${dataBaseLogin[$u]}" ] && [ -n "${dataBasePassword[$u]}" ] && [ -n "${backupProjectName[$u]}" ]
         then
           dbs=$(mysql -u${dataBaseLogin[$u]} -p${dataBasePassword[$u]} -e "show databases;" | grep [[:alnum:]])
 
           # The full path of the backup directory
-          pathway=${BACKUP_ROOT_DIR}"/"${backupProjectName[$i]}"/"${date_archived};
+          pathway="$BACKUP_ROOT_DIR/${backupProjectName[$u]}/$date_archived";
 
           # Check the existence of the backup directory
-          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check the existence of the backup directory for MySQL backup";
+          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check the existence of the backup directory for MySQL backup";
 
           if [ ! -d "${pathway}" ]
           then
             # Create a directory for the archive, if it was not created earlier
-            message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Create a directory for the archive, if it was not created earlier";
+            message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Create a directory for the archive, if it was not created earlier";
 
             mkdir -p ${pathway}
           fi;
 
           # Check if you need to archive all databases
-          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Check if you need to archive all databases";
+          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Check if you need to archive all databases";
 
           if [ "${ALL_DATA_BASE}" = "yes" ]
           then
@@ -225,7 +229,7 @@ create_backup()
             mv /tmp/$file ${pathway}"/sql/"${file}
           fi;
 
-		  message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Databases archiving is completed";
+          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Databases archiving is completed";
         fi;
 
         u=$(( u + 1 ))
@@ -241,25 +245,29 @@ clean_by_date ()
   then
     if [ "${MAX_NUMBER_DAYS}" -gt "0" ];
     then
-	  message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Start rotation of old copies, by date / time";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Start rotation of old copies, by date / time";
 
       # Check the project directories in turn
-      while [ "$k" != "${projects_counter}" ]; do
-        # Backup directory
-        pathway=${BACKUP_ROOT_DIR}"/"${backupProjectName[$k]}"/";
+      while [ "$k" != "${projects_counter}" ];
+      do
+        if [ -n ${backupProjectName[$k]} ];
+        then
+          # Backup directory
+          pathway="$BACKUP_ROOT_DIR/${backupProjectName[$k]}/$date_archived";
 
-        # Checking the nested directories
-        message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Checking the nested directories";
+          # Checking the nested directories
+          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Checking the nested directories";
 
-        for i in `ls ${pathway} -l -1t | grep '^d' |awk '{print $8}'`;
-        do
-          find ${pathway}${i} -mtime +${MAX_NUMBER_DAYS} -type d -exec rm -rf {} \;
-        done
+          for i in `ls ${pathway} -l -1t | grep '^d' |awk '{print $8}'`;
+          do
+            find ${pathway}${i} -mtime +${MAX_NUMBER_DAYS} -type d -exec rm -rf {} \;
+          done
 
-        k=$(( k + 1 ))
+          k=$(( k + 1 ))
+        fi;
       done
 
-	  message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Finish rotation of old copies, by date / time";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Finish rotation of old copies, by date / time";
     fi;
   fi;
 }
@@ -271,34 +279,36 @@ clean_by_count ()
   then
     if [ "${MAX_NUMBER_ARCHIVES}" -gt "0" ];
     then
-      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Start rotation of old copies, by number";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Start rotation of old copies, by number";
 
       # Check the project directories in turn
-      while [ "$k" != "${projects_counter}" ]; do
-        # Backup directory
-        pathway=${BACKUP_ROOT_DIR}"/"${backupProjectName[$k]}"/";
+      while [ "$k" != "${projects_counter}" ];
+      do
+        if [ -n ${backupProjectName[$k]} ];
+        then
+          # Backup directory
+          pathway="$BACKUP_ROOT_DIR/${backupProjectName[$k]}/";
 
-        preCount=$((MAX_NUMBER_ARCHIVES-1));
+          preCount=$((MAX_NUMBER_ARCHIVES-1));
 
-        # Checking the nested directories
-        message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Checking the nested directories";
+          # Checking the nested directories
+          message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Checking the nested directories";
 
-        echo "$(ls ${pathway} -l -1t | grep '^d' |awk '{print $9}')";
+          for i in $(ls ${pathway} -l -1t | grep '^d' |awk '{print $9}');
+          do
+            if [ "${subdir_counter}" -ge "${preCount}" ]
+            then
+              rm -rf ${pathway}${i};
+            fi;
 
-        for i in $(ls ${pathway} -l -1t | grep '^d' |awk '{print $9}');
-        do
-          if [ "${subdir_counter}" -ge "${preCount}" ]
-          then
-            rm -rf ${pathway}${i};
-          fi;
+            let subdir_counter=$((${subdir_counter} + 1));
+          done
 
-          let subdir_counter=$((${subdir_counter} + 1));
-        done
-
-        k=$(( k + 1 ))
+          k=$(( k + 1 ))
+        fi;
       done
 
-	  message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S)\r Finish rotation of old copies, by number";
+      message_str="$message_str\r\r$(date +%Y_%m_%d_%H_%M_%S) - Finish rotation of old copies, by number";
     fi;
   fi;
 }
